Move shared types to `safe-core-sdk-types`
