+++ packages/safe-core-sdk-types/src/types.ts
@@ -252,3 +252,51 @@ export interface MetaTransactionOptions {
   gasToken?: string
   isSponsored?: boolean
 }
+
+export type UserOperation = {
+  sender: string
+  nonce: string
+  initCode: string
+  callData: string
+  callGasLimit: bigint
+  verificationGasLimit: bigint
+  preVerificationGas: bigint
+  maxFeePerGas: bigint
+  maxPriorityFeePerGas: bigint
+  paymasterAndData: string
+  signature: string
+}
+
+export type SafeUserOperation = {
+  safe: string
+  nonce: bigint
+  initCode: string
+  callData: string
+  callGasLimit: bigint
+  verificationGasLimit: bigint
+  preVerificationGas: bigint
+  maxFeePerGas: bigint
+  maxPriorityFeePerGas: bigint
+  paymasterAndData: string
+  validAfter: number
+  validUntil: number
+  entryPoint: string
+}
+
+export type EstimateGasData = {
+  maxFeePerGas?: bigint
+  maxPriorityFeePerGas?: bigint
+  preVerificationGas?: bigint
+  verificationGasLimit?: bigint
+  callGasLimit?: bigint
+}
+
+export interface SafeOperation {
+  readonly data: SafeUserOperation
+  readonly signatures: Map<string, SafeSignature>
+  getSignature(signer: string): SafeSignature | undefined
+  addSignature(signature: SafeSignature): void
+  encodedSignatures(): string
+  addEstimations(estimations: EstimateGasData): void
+  toUserOperation(): UserOperation
+}
